base: use Align argument type
The 'Align' type is meant to be used as argument for aligned allocations
and replaces the former use of basic unsigned values. Typical argument
values AT_PAGE, AT_16_BYTES, AT_MWORD are now defined by the public API,
alleviating many former uses of 'get_page_size_log2()'.

Fixes #5660

+++ repos/base-foc/src/core/include/util.h
@@ -67,8 +67,8 @@ namespace Core {
 	inline addr_t trunc_page(addr_t addr) { return Foc::l4_trunc_page(addr); }
 	inline addr_t round_page(addr_t addr) { return Foc::l4_round_page(addr); }
 
-	constexpr size_t get_super_page_size()      { return L4_SUPERPAGESIZE;      }
-	constexpr size_t get_super_page_size_log2() { return L4_LOG2_SUPERPAGESIZE; }
+	constexpr size_t  get_super_page_size()      { return L4_SUPERPAGESIZE;      }
+	constexpr uint8_t get_super_page_size_log2() { return L4_LOG2_SUPERPAGESIZE; }
 
 	inline addr_t map_src_addr(addr_t core_local_addr, addr_t) {
 		return core_local_addr; }
